Move logic one level up
