+++ src/WhereConditions/WhereConditionsBaseDirective.php
@@ -115,15 +115,10 @@ abstract class WhereConditionsBaseDirective extends BaseDirective implements Arg
                     $relation,
                     function ($builder) use ($relation, $model, $condition): void {
                         if ($condition) {
-                            $relatedModel = $this->nestedRelatedModel($model, $relation);
-
                             $this->handleWhereConditions(
                                 $builder,
-                                $this->prefixConditionWithTableName(
-                                    $condition,
-                                    $relatedModel
-                                ),
-                                $relatedModel
+                                $condition,
+                                $this->nestedRelatedModel($model, $relation)
                             );
                         }
                     },
@@ -197,19 +192,6 @@ abstract class WhereConditionsBaseDirective extends BaseDirective implements Arg
         });
 
         return $relatedModel;
-    }
-
-    /**
-     * @param array<string, mixed> $condition
-     * @return array<string, mixed>
-     */
-    protected function prefixConditionWithTableName(array $condition, Model $model): array
-    {
-        if ($condition['column'] ?? null) {
-            $condition['column'] = $model->getTable().'.'.$condition['column'];
-        }
-
-        return $condition;
     }
 
     /**
