+++ src/WhereConditions/WhereHasConditionsDirective.php
@@ -2,6 +2,7 @@
 
 namespace Nuwave\Lighthouse\WhereConditions;
 
+use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Str;
 
 class WhereHasConditionsDirective extends WhereConditionsBaseDirective
@@ -56,7 +57,13 @@ SDL;
             $builder,
             $builder->getModel(),
             $this->getRelationName(),
-            $whereConditions
+            $this->prefixConditionWithTableName(
+                $whereConditions,
+                $this->nestedRelatedModel(
+                    $builder->getModel(),
+                    $this->getRelationName()
+                )
+            ),
         );
 
         return $builder;
@@ -83,5 +90,18 @@ SDL;
     protected function generatedInputSuffix(): string
     {
         return 'WhereHasConditions';
+    }
+
+    /**
+     * @param array<string, mixed> $condition
+     * @return array<string, mixed>
+     */
+    protected function prefixConditionWithTableName(array $condition, Model $model): array
+    {
+        if ($condition['column'] ?? null) {
+            $condition['column'] = $model->getTable().'.'.$condition['column'];
+        }
+
+        return $condition;
     }
 }
